GameINI: Disable Fast Texture Sampling for EA games that use VP6 videos
This is based on the list at https://wiki.dolphin-emu.org/index.php?title=Template:Problems/VP6_Videos.  Per that template, Force Texture Filtering and Anisotropic Filtering have also been disabled.  Anisotropic Filtering being disabled does nothing currently, as Manual Texture Sampling does not implement it, but there is a chance that it will cause problems when it is implemented (this will need to be retested in the future).