PI_RESET_CODE attempt
This doesn't work because the write to PC by BootUp is ignored.  In the past it worked if the user sets the PC to 81200150 in the register view (things aren't happy when you do that, but it sorta works), but now setting the PC to fff00100 doesn't work (probably due to virtual vs physical addresses).